Move notes-saving code into NoteItem
