+++ visitz/Visitz/Services/GetNotesService.cs
@@ -1,4 +1,4 @@
-﻿using Visitz.Services.Messages;
+using Visitz.Services.Messages;
 using Visitz.Storage;
 using VisitzApi;
 using VisitzModel.Models;
@@ -48,20 +48,9 @@ namespace Visitz.Services
             var newNotes = NoteItem.FromApiEntities(id, notesFromApi);
 
             using var realm = await VisitzRealms.GetIcmDataRealmAsync();
-            var currentNotes = NoteItem.GetNotesByEntityId(realm, id);
-            var deletedNotes = currentNotes.ExceptBy(newNotes.Select(NoteSelector), NoteSelector);
-
-            await realm.WriteAsync(() =>
-            {
-                foreach (var deletedNote in deletedNotes)
-                    realm.Remove(deletedNote);
-
-                realm.Add(newNotes, update: true);
-            });
+			await NoteItem.UpsertNotesAsync(realm, id, newNotes);
 
             ResultCode = Result.Successful;
         }
-
-        static string NoteSelector(NoteItem note) => note.FullID;
     }
 }
